Pull ColdObservable out into its own module.
