+++ test/support/marble_testing.ex
@@ -3,11 +3,6 @@ defmodule MarbleTesting do
 
   alias VirtualTimeScheduler, as: VTS
 
-  defmodule ColdObservable do
-    @moduledoc false
-    defstruct [:events]
-  end
-
   @doc ~S"""
   Creates a cold observable for use in marble testing.
 
@@ -33,38 +28,14 @@ defmodule MarbleTesting do
   TODO: Change this so it runs core Observable type, not ColdObservable.
   """
   def observe(%__MODULE__.ColdObservable{} = observable) do
-    {r_notifs, subscriptions} = VTS.run(&subscribe/3, observable, {[], %{}})
+    # TODO: Wrap MT.CO.subscribe.
+    {r_notifs, subscriptions} = VTS.run(&MarbleTesting.ColdObservable.subscribe/3,
+                                        observable, {[], %{}})
     {Enum.reverse(r_notifs), subscriptions}
   end
 
   # TODO: Move this out to real subscription module. (I think.)
 
-  defp subscribe(time,
-                 %__MODULE__.ColdObservable{events: events} = observable,
-                 {r_notifs, subscriptions} = _acc)
-  do
-    # TODO: Generalize into a subscription function that all observables can impl.
-    # TODO: Figure out how to record unsubscription cleanly.
-
-    subscriptions = Map.put(subscriptions, observable, {time, nil})
-    {{r_notifs, subscriptions}, new_events: Enum.map(events, &schedule_event/1)}
-  end
-
-  defp schedule_event({time, :next, value}), do: {time, &do_next_event/3, value}
-  defp schedule_event({time, :done}), do: {time, &do_done_event/3, nil}
-
-  # TODO: Generalize into a subscription wrapper/runner that calls Observable
-  # and marshals these responses appropriately.
-  defp do_next_event(time, value, {r_notifs, subscriptions} = _acc) do
-    {{[{time, :next, value} | r_notifs], subscriptions}, []}
-  end
-  defp do_done_event(time, nil, {r_notifs, subscriptions} = _acc) do
-    # NASTY HACK: There can be only one subscription, right?
-    the_one_true_observable = List.first(Map.keys(subscriptions))
-    {sub, nil} = subscriptions[the_one_true_observable]
-    subscriptions = Map.put(subscriptions, the_one_true_observable, {sub, time})
-    {{[{time, :done} | r_notifs], subscriptions}, []}
-  end
 
   @doc ~S"""
   Converts a string containing a marble diagram of notifications into a sequence
